+++ spec/scrub/member_holding_spec.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require "loader/member_holding"
+require "scrub/member_holding"
 require "custom_errors"
 require "json"
 
-RSpec.describe Loader::MemberHolding do
+RSpec.describe Scrub::MemberHolding do
   let(:test_data) { __dir__ + "/../testdata" }
   let(:ok_min_hed) { { "oclc"=>0, "local_id"=>1 } }
   let(:ok_min_str) { "123\t456" }
